move backend tests to single file
